changed int to Float for hours

+++ app/src/main/java/com/georgcantor/counter/model/Day.kt
@@ -7,5 +7,5 @@ import androidx.room.PrimaryKey
 data class Day(
     @PrimaryKey
     val id: String,
-    val hours: Int
+    val hours: Float
 )
